mtp_cmd_handler: fix possible inotify raise
At the moment the inotify thread can get triggered in parallel to the
mtp_cmd_handler thread due to the _util_file_copy() call which copies
the tmpfile (containing the received data) to the data-store. Since the
inotify execution behaviour depends on the g_is_send_* variables we need
to set them before.

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>