mtp_cmd_handler: fix PTP_OC_ANDROID_SENDPARTIALOBJECT
The SendPartialObject was broken in the way the src/dst path was
handled. The g_copy_src_file is only updated during the GetPartialObject
cmd which is not involved during sending updates from the host to the
device.

For g_copy_dst_file it is a bit more complex. The path would be correct
if the file is newly created due to some side effects of the inotify
thread which sets the g_copy_dst_file correctly. But this mechanism is
not working if someone updates an exising file. In such case the content
is written to an tmpfile under /tmp and inotify is not triggered if the
storage is not set to /tmp. In this case the g_copy_dst_file is empty
and the _util_file_copy() fail.

The fix is to use the ftemp_st.filepath as src and obj->file_path as
dst.

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>