main: fix existing file update host propagation
Currently the file update will not be propagated to the host if the file
already exists. In this case the inotify IN_CLOSE_WRITE is triggered but
the edited file wasn't added to the open_files_list. At the moment only
new files are added to this list while handling IN_CREATE.

Fix this by adding the IN_MODIFY to the list of listen inotify events.
If the event is triggered it adds the file to the open_files_list so the
IN_CLOSE_WRITE handling can find the file and update the store
accordingly which triggeres an host notify in the end.

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>